Touch Event(s) + Basic Scoring and "Gameplay"
Loads into a screen with a Pentagon and a Score in the Upper Left corner
of the Screen. Clicking on the pentagon will reward points based on how
short the interval is between your last click and your latest successful
click.

I realize this isn't exactly the Gameplay we talked about, but I just
iterated into it as I was testing things.